restrict password verification to the caller's own account
